some fixBugs replace buttonJs move button/components
