+++ src/pages/Login/Login.css
@@ -1,4 +1,4 @@
-.Login-page {
+.LoginPage {
   height: 100vh;
   background: #44B26F;
   display: flex;
