+++ src/pages/Login/components/Card/Card.css
@@ -1,0 +1,14 @@
+
+.Card {
+  background: #f5f5f5;
+  padding: 80px 105px;
+  padding-bottom: 200px;
+}
+
+.loginLogo {
+  margin-bottom: 85px;
+}
+
+.inputLogin {
+  margin-bottom: 46px;
+}
