+++ src/pages/ProductsPage/components/ProductsTable/ProductsTable.js
@@ -61,9 +61,6 @@ class ProductsTable extends Component {
           </div>
 
 
-
-
-
         )
       }
     }
